230620 Only owner can edit thier profile and video

+++ src/controllers/videoController.js
@@ -1,3 +1,4 @@
+import User from "../models/User";
 import Video from "../models/video";
 
 /*
@@ -19,7 +20,7 @@ export const home = async(req, res) => {
 };
 export const watch = async (req, res) => {
     const { id } = req.params;
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate("owner");
     if(video){
         return res.render("watch", { pageTitle: video.title, video });
     }
@@ -28,23 +29,31 @@ export const watch = async (req, res) => {
 
 export const getEdit = async (req, res) => {
     const { id } = req.params;
+    const { user: { _id }, } = req.session;
     const video = await Video.findById(id);
     if (!video) {
         return res. render("404", { pageTitle: "Video not found."});
         // return을 넣지 않으면 계속해서 밑에있는 코드들도 실행되기때문에 return을 넣어서 끝내줘야한다
     }
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    }
     return res.render("edit", { pageTitle:`Edit: ${video.title}`, video });
 }
 
 export const postEdit = async (req, res) => {
+    const { user: { _id }, } = req.session;
     const { id } = req.params;
     const { title, description, hashtags } = req.body;
     const video = await Video.exists({ _id: id });
     if (!video) {
         return res. render("404", { pageTitle: "Video not found."});
     }
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    }
     await Video.findByIdAndUpdate(id, {
-        title, discription, hashtags: Video.formatHashtags(hashtags),
+        title, description, hashtags: Video.formatHashtags(hashtags),
     });
 
     return res.redirect(`/videos/${id}`);
@@ -55,13 +64,21 @@ export const getUpload = (req, res) => {
 };
 
 export const postUpload = async (req, res) => {
+    const { user: {_id},} = req.session;
+    const { path: fileUrl } = req.file;
     const { title, description, hashtags } = req.body;
     try {
-        await video.create({
+        const newVideo = await video.create({
         title: title,
         description: description,
+        fileUrl,
+        // es6를 공부하면 위처럼 코드를 짧게 줄일 수 있대
+        owner: _id,
         hashtags: Video.formatHashtags(hashtags),
         });
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id);
+        user.save();
     return res.redirect("/");
     } catch(error){
         return res.status(400).render("upload", { pageTitle: "Upload Video", errorMessage: error._message, })};
@@ -70,6 +87,14 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
     const { id } = req.params;
+    const { user: { _id }, } = req.session;
+    const video = await Video.findById(id);
+    if (!video) {
+        return res.status(404).render("404", { pageTitle: "Video not found."});
+    }
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    }
     await Video.findByIdAndDelete(id);
     return res.redirect("/");
 };
